+++ packages/create-content-sdk-app/src/templates/nextjs-app-router (beta)/src/middleware.ts
@@ -66,13 +66,14 @@ const personalize = new PersonalizeMiddleware({
   // This is an important performance consideration since Next.js Edge middleware runs on every request.
   skip: () => false,
   // This is an example of how to provide geo data for personalization.
-  // This has to set per per request, as geo data varies between users.
-  // In real world scenarios, you would extract geo data from a 3rd party geo service or your own service.
-  // geo:{
-  //   city: 'Athens',
-  //   country: 'Greece',
-  //   region: 'Attica',
-  // }
+  // The provided callback will be called on each request to extract geo data.
+  // extractGeoDataCb: () => {
+  //   return {
+  //     city: 'Athens',
+  //     country: 'Greece',
+  //     region: 'Attica',
+  //   };
+  // },
 });
 
 export function middleware(req: NextRequest, ev: NextFetchEvent) {