+++ packages/create-content-sdk-app/src/templates/nextjs/src/middleware.ts
@@ -18,8 +18,8 @@ export function middleware(req: NextRequest, ev: NextFetchEvent) {
   // Instantiate AFTER the guard so constructors don’t run in local-only mode
   const multisite = new MultisiteMiddleware({
     /**
-    * List of sites for site resolver to work with
-    */
+     * List of sites for site resolver to work with
+     */
     sites,
     ...scConfig.api.edge,
     ...scConfig.multisite,
@@ -31,8 +31,8 @@ export function middleware(req: NextRequest, ev: NextFetchEvent) {
 
   const redirects = new RedirectsMiddleware({
     /**
-    * List of sites for site resolver to work with
-    */
+     * List of sites for site resolver to work with
+     */
     sites,
     ...scConfig.api.edge,
     ...scConfig.redirects,
@@ -45,8 +45,8 @@ export function middleware(req: NextRequest, ev: NextFetchEvent) {
 
   const personalize = new PersonalizeMiddleware({
     /**
-    * List of sites for site resolver to work with
-    */
+     * List of sites for site resolver to work with
+     */
     sites,
     ...scConfig.api.edge,
     ...scConfig.personalize,
@@ -56,13 +56,14 @@ export function middleware(req: NextRequest, ev: NextFetchEvent) {
     // This is an important performance consideration since Next.js Edge middleware runs on every request
     skip: () => false,
     // This is an example of how to provide geo data for personalization.
-    // This has to be set per request, as geo data varies between users.
-    // In real world scenarios, you would extract geo data from a 3rd party geo service or your own service.
-    // geo:{
-    //   city: 'Athens',
-    //   country: 'Greece',
-    //   region: 'Attica',
-    // }
+    // The provided callback will be called on each request to extract geo data.
+    // extractGeoDataCb: () => {
+    //   return {
+    //     city: 'Athens',
+    //     country: 'Greece',
+    //     region: 'Attica',
+    //   };
+    // },
   });
 
   return defineMiddleware(multisite, redirects, personalize).exec(req, ev);