+++ packages/nextjs/src/middleware/personalize-middleware.test.ts
@@ -178,6 +178,7 @@ describe('PersonalizeMiddleware', () => {
       personalizeStub?: sinon.SinonStub;
       handleCookieStub?: sinon.SinonStub;
       getClientFactoryStub?: sinon.SinonStub;
+      extractGeoDataCb?: sinon.SinonStub;
     } = { config: defaultConfig }
   ) => {
     const clientFactory = GraphQLRequestClient.createClientFactory({
@@ -1056,6 +1057,7 @@ describe('PersonalizeMiddleware', () => {
     });
 
     describe('geo data', () => {
+      const geo = { country: 'US', region: 'CA', city: 'San Francisco' };
       const req = createRequest();
       const res = createResponse();
       const personalizeInfo = {
@@ -1075,10 +1077,40 @@ describe('PersonalizeMiddleware', () => {
       });
 
       it('should call personalize with geo data', async () => {
-        const geo = { country: 'US', region: 'CA', city: 'San Francisco' };
+        const extractGeoDataCb = sandbox.stub().returns(geo);
 
         const { middleware, initPersonalizeServer } = createMiddleware({
+          personalizeInfo,
+          extractGeoDataCb,
+        });
+
+        middleware['personalize'] = PersonalizeMiddleware.prototype['personalize'];
+
+        await middleware.handle(req, res);
+
+        validateDebugLog('personalize middleware start: %o', {
           geo,
+          headers: {
+            ...req.headers,
+          },
+          hostname: 'foo.net',
+          pathname: '/styleguide',
+          language: 'en',
+        });
+
+        expect(extractGeoDataCb.calledOnce).to.be.true;
+        expect(initPersonalizeServer.calledOnce).to.be.true;
+        expect(CDKPersonalizeStub.calledThrice).to.be.true;
+        expect(CDKPersonalizeStub.firstCall.args[1].geo).to.deep.equal(geo);
+        expect(CDKPersonalizeStub.secondCall.args[1].geo).to.deep.equal(geo);
+        expect(CDKPersonalizeStub.thirdCall.args[1].geo).to.deep.equal(geo);
+      });
+
+      it('should call personalize with geo data when an async cb is provided', async () => {
+        const extractGeoDataCb = sandbox.stub().resolves(geo);
+
+        const { middleware, initPersonalizeServer } = createMiddleware({
+          extractGeoDataCb,
           personalizeInfo,
         });
 
@@ -1095,6 +1127,8 @@ describe('PersonalizeMiddleware', () => {
           pathname: '/styleguide',
           language: 'en',
         });
+
+        expect(extractGeoDataCb.calledOnce).to.be.true;
         expect(initPersonalizeServer.calledOnce).to.be.true;
         expect(CDKPersonalizeStub.calledThrice).to.be.true;
         expect(CDKPersonalizeStub.firstCall.args[1].geo).to.deep.equal(geo);